Action.js: Avoid JSON parse errors
In Chrome and Firefox, `innerHTML` we get from the backend is
wrapped in (different) forms of `<pre></pre>`.
(Chrome has additional inline styling.)

Remove these tags before trying to parse it as JSON.

Also, parsing an empty string is an error, which still got us the standard "9999 No Data" error.